% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculateDominance.R
\name{calculateDominance}
\alias{calculateDominance}
\title{Calculates the Dominance Value for a given tuple of coordinates and altitude
as derived from DEM and provided by the peaklist}
\usage{
calculateDominance(x.coord, y.coord, altitude)
}
\arguments{
\item{altitude}{altitude in meter}

\item{xcoord}{xcoordinate (mapunits)}

\item{ycoord}{ycoordinate (mapunits)}
}
\value{
calculateDominance returns the dominance value of the current peak
}
\description{
Calculates the Dominance Value for a given tuple of coordinates and altitude
as derived from DEM d = dominance horizontal distance to the next higher altirude in the sourrounding
Sources:    Rauch. C. (2012): Der perfekte Gipfel.  Panorama, 2/2012, S. 112
            http://www.alpenverein.de/dav-services/panorama-magazin/dominanz-prominenz-eigenstaendigkeit-eines-berges_aid_11186.html#             (Zugriff: 20.05.2012)
            Leonhard, W. (2012): Eigenständigkeit von Gipfeln. -
}
\examples{
#### Example to parse an windows type of ini file
#### create a list for each ini section containing the variables and params

calculateDominance <- function(peaklist.tupel)
}
\author{
Chris Reudenbach
}
\references{
Marburg Open Courseware Advanced GIS: \url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}

Rauch. C. (2012): Der perfekte Gipfel.  Panorama, 2/2012, S. 112 \url{http://www.alpenverein.de/dav-services/panorama-magazin/dominanz-prominenz-eigenstaendigkeit-eines-berges_aid_11186.html}

Leonhard, W. (2012): Eigenständigkeit von Gipfeln.\url{http://www.thehighrisepages.de/bergtouren/na_orogr.htm}
}

