% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eValue.R
\name{calculateEValue}
\alias{calculateEValue}
\title{Calculates the Eigenstaendigkeitswert of a summit with given parameters dominance (d) and prominence (p)}
\usage{
calculateEValue(peaklist.tupel)
}
\arguments{
\item{peaklist.tupel}{one row of the peaklist, containing all the necessary values}
}
\value{
calculateEValue returns the E Value of the current peak
}
\description{
Calculates the independence Value (E) for a given tuple of coordinates and
altitude as derived from DEM According to the formula as suggested by Rauch
if d<100000: E = -((log2  (h / 8848) + log2 (d / 100000) + log2(p / h)) / 3)
if d>100000: E = -((log2 (h / 8848) + log2(p / h)) / 3)
There is a controverse discussion of the formula at:
http://www.alpenverein.de/dav-services/panorama-magazin/dominanz-prominenz-eigenstaendigkeit-eines-berges_aid_11186.html
}
\details{
\tabular{ll}{
Package: \tab Rpeak\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\examples{
#### Example to parse an windows type of ini file
#### create a list for each ini section containing the variables and params

calculateEValue <- function(peaklist.tupel)
}
\author{
Chris Reudenbach
}
\references{
\url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}
}

