% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateProminence.R
\name{calculateProminence}
\alias{calculateProminence}
\title{Calculates the prominence Value for a given tuple of coordinates and altitude}
\usage{
calculateProminence(x.coord, y.coord, altitude)
}
\arguments{
\item{peaks}{list of all peaks}

\item{altitude}{altitude in meter}

\item{exact.enough}{= vertical excactness of notch altitude in meter}

\item{xcoord}{xcoordinate of current peak (mapunits)}

\item{ycoord}{ycoordinate of current peak (mapunits)}
}
\value{
calculateProminence returns the prominence value of the current peak
}
\description{
Calculates the prominence Value for a given tuple of coordinates and altitude
as derived from DEM d = dominance horizontal distance to the next higher altirude in the sourrounding
Sources:    Rauch. C. (2012): Der perfekte Gipfel.  Panorama, 2/2012, S. 112
            http://www.alpenverein.de/dav-services/panorama-magazin/dominanz-prominenz-eigenstaendigkeit-eines-berges_aid_11186.html#             (Zugriff: 20.05.2012)
            Leonhard, W. (2012): Eigenständigkeit von Gipfeln. -
}
\details{
\tabular{ll}{
Package: \tab Rpeak\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\examples{
#### Example to calculateProminence

calculateProminence(peaklist.tupel)
}
\author{
Chris Reudenbach
}
\references{
\url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}
}

