% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/costMergePeaks.R
\name{costmergePeaks}
\alias{costMergePeaks}
\alias{costmergePeaks}
\title{Trys to merge indenpendent peak positions to provide a reliable join using a cost analysis}
\usage{
costmergePeaks(dem.peaklist,ext.peaklist)
}
\arguments{
\item{dem.peaklist}{DEM derived peaklist containing coords and altitude}

\item{ext.peaklist}{external peaklist, containing coords name and altitude}

\item{dem}{DEM data as a raster object}
}
\value{
costmergePeaks returns one merged peaklist with the names of the external peaks and the coords and altitude from the DEM derived list
}
\description{
http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description
}
\details{
\tabular{ll}{
Package: \tab Rpeak\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\examples{
#### Example merge two peak lists
      new.peaklist<-costmergePeaks(dem.peaklist,ext.peaklist)
}
\author{
Chris Reudenbach
}
\references{
\url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}
}

