% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extractOSMPoints.R
\name{extractOSMPoints}
\alias{extractOSMPoints}
\title{Extract peak position and altitude from OSM data}
\usage{
extractOSMPoints(dem,current.proj4)
}
\arguments{
\item{dem}{is a Digital Elevation Model with geographic coordinates}

\item{target.projection}{is a valid proj.4 string containing the correct target crs}
}
\value{
extractOSMPoints returns the following parameters:
\tabular{ll}{
xcoord \tab xcoordinate (mapunits)\cr
ycoord \tab ycoordinate (mapunits)\cr
altitude \tab altitude in meter\cr
name \tab name of the peak\cr
}
}
\description{
The current OSM data base is cropped for th earea of interest Kmz file of Harry's peak list is downloaded and will be cleaned to derive coordinates altitude and name of all available peaks within a region of interest
}
\details{
\tabular{ll}{
Package: \tab Rpeak\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\examples{
#### Example extract OSM nodes from a region of interst (ROI)
#    create a dataframe containing coordinates, tag-attribute and value
ini.example=system.file("data","demo.ini", package="perfectPeak")
dem.example=system.file("data","test.asc", package="perfectPeak")
ini<-initEnvironGIS(ini.example,dem.example)
key="natural"
val="peak"
extractOSMPoints(ini,k,v)
}
\author{
Chris Reudenbach
}
\references{
\url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}
}

