% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfectPeak-package.R
\docType{data}
\name{INI}
\alias{INI}
\title{Demo Ini File}
\format{Windows style INI file}
\usage{
iniparse('demo.ini')
}
\description{
The example control file provides all necessary settings for using the perfectPeak package
\tabular{ll}{
[Pathes]\tab\cr
workhome='/home/MOC/aGIS/Rpeak' \tab # project root directory\cr
rawdata='./data' \tab # data directrory relative to 'workhome'\cr
runtimedata='run' \tab # runtime directory relative to 'workhome'\cr
src='./src' \tab # script directory relative to 'workhome'\cr
}
\tabular{ll}{
[Files]\tab\cr
peaklist='peaklist.txt' \tab    # output fname for ASCII peaklist\cr
fndem='stubai.asc' \tab         #  input \link{DEM} (has to be GDAL conform)\cr
}
\tabular{ll}{
[Projection] \tab\cr
targetepsg='31254' \tab # EPSG code for Austian DEM data NOTE you have to use the targetproj4 string for all projections  \cr
targetproj4='+proj=tmerc +lat_0=0 +lon_0=10.33333333333333 +k=1 +x_0=0 +y_0=-5000000 +ellps=bessel +towgs84=577.326,90.129,463.919,5.137,1.474,5.297,2.4232 +units=m +no_defs' \tab # in this case the CORRECT proj4 string for Austrian DEM data set, usually the corresponding proj4 string\cr
latlonproj4='+proj=longlat +datum=WGS84 +no_defs' \tab # common geographic coordinates =EPSG:4326 \cr
}
\tabular{ll}{
[Params]\tab\cr
makepeakmode=3 \tab # 1=simple local minmax, 2=wood 3=fuzzylandforms\cr
filterkernelsize=9 \tab # size of filter for makepeak=1; range 3-30; default=9 \cr
externalpeaks='osm' \tab # 'harry'= Harrys Peaklist \link{PeakList} osm= OSM peak data\cr
mergemode=1 \tab # merging peak names from harry/OSM/user data 1=distance (default) 2=costpath (slow)\cr
exactenough=5 \tab # treshold of vertical matching exactness (meter) during prominence analysis\cr
domthres=100 \tab # threshold of minimum dominance distance (makepeak=2&3)\cr
}
[Wood] # Wood parameters refers to SAGA\cr
WSIZE=11\cr
TOL_SLOPE=14.000000\cr
TOL_CURVE=0.00001\cr
EXPONENT=0.000000\cr
ZSCALE=1.000000\cr
\cr
[FuzzyLf] # FuzzyLandforms parameters refers to SAGA\cr
SLOPETODEG=0\cr
T_SLOPE_MIN=0.0000001\cr
T_SLOPE_MAX=25.000000\cr
T_CURVE_MIN=0.00000001\cr
T_CURVE_MAX=0.001\cr
\cr
# The below SYSPath for Linux/Windows are correct according to \url{http://giswerk.org/doku.php?id=projekte:gis-software:gis-distros}\cr
\cr 
[SysPath]            
wossaga='C:/MyApps/GIS_ToGo/QGIS_portable_Chugiak_24_32bit/QGIS/apps/saga'\cr
wsagamodules='C:/MyApps/GIS_ToGo/QGIS_portable_Chugiak_24_32bit/QGIS/apps/saga/modules'\cr
wgrassgisbase='C:/MyApps/GIS_ToGo/GRASSGIS643/bin'\cr
lossaga='/home/creu/SAGA-2.1.0-fixed/initial/bin'\cr
lsagamodules='/home/creu/SAGA-2.1.0-fixed/initial/lib/saga'\cr
lgrassgisbase='/usr/lib/grass64'\cr
}
\keyword{datasets}

