% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/initParse.R
\name{iniParse}
\alias{iniParse}
\title{Parse an Windows like ini file to get the the entries}
\usage{
iniParse(fname)
}
\arguments{
\item{fname}{file name of the ASCII INI file}
}
\value{
iniParse returns a list ordered by sections variablenames and values
}
\description{
Any Windows type ini file will be parsed so that the sections and the variables may be used within R for all kind of settings.
}
\examples{
#### Example to parse an windows type of ini file
#### create a list for each ini section containing the variables and params

ini.example=system.file("demo.ini", package="Rpeak")
iniParse(ini.example)
}
\author{
Gabor Grothendieck <ggrothendieck at gmail.com>,
\cr
\emph{Maintainer:} Chris Reudenbach \email{giswerk@gis-ma.org}
}
\references{
\url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}
           \url{https://stat.ethz.ch/pipermail/r-help/2007-June/134115.html}
}

