% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initEnvironGIS.R
\name{initEnvironGIS}
\alias{initEnvironGIS}
\title{Function that initializes RSAGA, GRASS GDAL and the R packages}
\usage{
initEnvironGIS(fname.control)
}
\arguments{
\item{fname.control}{name of the session ini.file}

\item{fname.DEM}{name of used raster DEM (GDAL format)}
}
\value{
initEnvironGIS initializes the usage of the GIS packages and other utilities
}
\description{
Function that initializes environment pathes, SAGA, GRASS7 and GDAL support.
            *NOTE* you probably have to customize some settings in your custom or the demo.ini file
}
\details{
This concept is somehow not very R-like. You have to provide and control at least the installations pathes of the used SAGA and GRASS executables, root and working folder, epsg code AND correct corresponding proj4 string of the DEM data file (all controlled by the ini file). Finally you need to provide a GDAL conform DEM file.

For Further Information look at:\link{INI}
}
\examples{
  
#### Example to initialize the enviroment and GIS bindings for use with R
#### uses the ini list from an ini file
      
ini.example=system.file("data","demo.ini", package="perfectPeak")
dem.example=system.file("data","deḿo.asc", package="perfectPeak")
initEnvironGIS(ini.example,dem.example)
gmeta()

}
\author{
Chris Reudenbach
}

