% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rpeak.R
\docType{data}
\name{input.DEM}
\alias{input.DEM}
\title{Example Digital Elevation Model DEM}
\format{Arc ASCII Grid}
\source{
Data source: \url{https://www.tirol.gv.at/data/datenkatalog/geographie-und-planung/digitales-gelaendemodell-tirol/}
}
\usage{
raster('stubai.asc')
}
\description{
The example Digital Elevation Model DEM is taken from the Authority of Tirol
it is dervied from LIDAR data and can downloaded for Tirol. The demo data is a tile
with  the following extent:  xmin 11.35547, xmax 11.40009, ymin 47.10114, ymax 47.13512
}
\details{
\itemize{
  \item resolution : 10 Meter,10 Meter
  \item datatype, 32 Bit floating point
  \item projection, MGI_Austria_GK_West
  \item EPSG-Code, 31254
  \item unit, Meter
  \item datum, D_MGI
  \item Copyright:   \url{https://www.tirol.gv.at/data/nutzungsbedingungen/},  \url{Creative Commons Namensnennung 3.0 Österreich Lizenz (CC BY 3.0 AT).}
  }
}
\keyword{datasets}

