% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rpeak.R
\docType{data}
\name{input.INI}
\alias{input.INI}
\title{Demo Ini File}
\format{Windows style INI file}
\usage{
iniparse('demo.ini')
}
\description{
The example control file provides all necessary settings for using the perfectPeak package
}
\details{
\itemize{
  \item [Pathes]
  \item workhome='/home/creu/MOC/aGIS/Rpeak'
  \item rawdata='./data'
  \item runtimedata='run'
  \item src='./src'
}
\itemize{
  \item [Files]
  \item peaklist='peaklist.txt'       output fname for ASCII peaklist
  \item fndem='stubai.asc'            input DEM (has to be GDAL conform)
}
\itemize{
  \item [Projection]
  \item targetepsg='31254'             epsg code of DEM data
  \item targetproj4='+proj=tmerc +lat_0=0 +lon_0=10.33333333333333 +k=1 +x_0=0 +y_0=-5000000 +ellps=bessel +towgs84=577.326,90.129,463.919,5.137,1.474,5.297,2.4232 +units=m +no_defs'
  \item latlonproj4='+proj=longlat +datum=WGS84 +no_defs'
}
\itemize{
  \item [Params]
  \item makepeakmode=3        1=simple local minmax, 2=wood 3=fuzzylandforms
  \item filterkernelsize=9    size of filter for makepeak=1; range 3-30; default=9
  \item externalpeaks='osm'   harry= Harrys Peaklist osm= OSM peak data
  \item mergemode=1           merging peak names from harry/OSM/user data 1=distance 2=costpath (slow)
  \item exactenough=5         prominnence treshold of vertical matching exactness (m)
  \item domthres=100          threshold of minimum dominance distance (makepeak=2&3)
}
\itemize{
  \item [Wood]               Wood parameters refers to SAGA
  \item WSIZE=11
  \item TOL_SLOPE=14.000000
  \item TOL_CURVE=0.00001
  \item EXPONENT=0.000000
  \item ZSCALE=1.000000
}
\itemize{
  \item [FuzzyLf]            FuzzyLandforms parameters refers to SAGA
  \item SLOPETODEG=0
  \item T_SLOPE_MIN=0.0000001
  \item T_SLOPE_MAX=25.000000
  \item T_CURVE_MIN=0.00000001
  \item T_CURVE_MAX=0.001
}

\itemize{
  \item [SysPath]             SYSPAth for Linux and Windows
  \item wossaga='C:/MyApps/GIS_ToGo/QGIS_portable_Chugiak_24_32bit/QGIS/apps/saga'
  \item wsagamodules='C:/MyApps/GIS_ToGo/QGIS_portable_Chugiak_24_32bit/QGIS/apps/saga/modules'
  \item wgrassgisbase='C:/MyApps/GIS_ToGo/GRASSGIS643/bin'
  \item lossaga='/home/creu/SAGA-2.1.0-fixed/initial/bin'
  \item lsagamodules='/home/creu/SAGA-2.1.0-fixed/initial/lib/saga'
  \item lgrassgisbase='/usr/lib/grass64'
  }
}
\keyword{datasets}

