% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/perfectPeak-package.R
\docType{data}
\name{PeakList}
\alias{PeakList}
\title{Harry's Peak List}
\format{KMZ File}
\source{
Data source: \url{http://www.tourenwelt.info/bergliste/bergliste.php}
}
\usage{
getGeoData('harrylist')
}
\description{
Harry's mountain list is a collection of peaks in the world. In the database are 58532 entries at the moment (15.1.2014). This is equivalent to 52218 summits (6314 aliases) in different mountain ranges all over the world.
}
\details{
EPSG-Code: 4326

Copyright:   \url{http://www.tourenwelt.info/impressum.php}
}
\keyword{datasets}

