\name{perfectPeak}

\docType{package}
\title{
Title: Analyse digitial elevation models (DEM) and derives dominance, prominence and independence values to classify perfect peaks
}
\description{
Organises all necessary processing steps for calculating the
perfect peak  parameters and generating the output. It  performs preprocessing
and controls the calculations of dominance, prominence, independence (E) value
for a given georeferencend Digital Elevation Model (DEM). additionally you may use external data as OSM data to name the derived peak entities.


}
\details{
\tabular{ll}{
Package: \tab Rpeak\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2015-01-19\cr
Date: \tab \Sexpr[echo=TRUE]{paste0(getYear(Sys.Date()),"-",getMonth(Sys.Date()),"-",getDay(Sys.Date()))}\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
#'}
}

\author{
Chris Reudenbach

Maintainer: Chris Reudenbach <creuden@@gmail.com>
}
\references{
http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description
}


\keyword{ package,prominence,dominance,summit,peak,DEM }





