% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfectPeak-package.R
\docType{package}
\name{perfectPeak}
\alias{perfectPeak}
\alias{perfectPeak-package}
\title{perfectPeak}
\description{
perfectPeak
}
\details{
A collection of functions to facilitate the processing steps for calculating the 
"perfect peak"  parameters and generating the output. There are functions for 
preprocessing and retrieving the values of dominance, prominence, independence  
from a given georeferencend Digital Elevation Model (DEM).
}
\author{
Chris Reudenbach\cr
\cr
\emph{Maintainer:} Chris Reudenbach \email{giswerk@gis-ma.org}
}
\references{
Marburg Open Courseware Advanced GIS: \url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}\cr
Rauch. C. (2012): Der perfekte Gipfel.  Panorama, 2/2012, S. 112 \url{http://www.alpenverein.de/dav-services/panorama-magazin/dominanz-prominenz-eigenstaendigkeit-eines-berges_aid_11186.html}\cr
Leonhard, W. (2012): Eigenständigkeit von Gipfeln.\url{http://www.thehighrisepages.de/bergtouren/na_orogr.htm}\cr
}
\seealso{
\pkg{perfectPeak} is using heavily SAGA and GRASS GIS via the \code{\link{RSAGA-package}} and 
 \code{\link{rgrass7-package}}.
Please see their documentation for special questions dealing with the algorithms and parameters etc.
}
\keyword{package}

