% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rpeak.R
\name{Rpeak}
\alias{Rpeak}
\title{Example script that can be used as a wrapper function to start the
perfect peak analysis}
\usage{
Rpeak(fname.control,fname.DEM)
}
\arguments{
\item{fname.control}{name of control file containing all setting and parameters for analysis}

\item{fname.DEM}{name Digtial Elevation Model has to be a GDAL raster file}
}
\value{
Rpeak returns the complete list as a dataframe of all parameters and results and
generates some output (maps and tables)
}
\description{
Organises all necessary processing steps for calculating the
perfect peak  parameters and generating the output. It  performs preprocessing
and controls the calculations of dominance, prominence, independence (E) value
for a given georeferencend Digital Elevation Model (DEM)

You can use the function as it is or alternatively use it as skeleton control
script that you cab adapt to your needs.
}
\examples{
#### Example to use Rpeak for a common analysis run

# You obligatory need a georeferenced DEM (GDAL format) as data input.
Except the origin georefence all parameters are read from the demo.ini
file. You will find some more comments in the file. NOTE the real projection of
the DEM has to meet the projection string in the ini file


ini.example=system.file("data","demo.ini", package="perfectPeak")
dem.example=system.file("data","test.asc", package="perfectPeak")
Rpeak(ini.example,dem.example)
}
\author{
Chris Reudenbach
}
\references{
Marburg Open Courseware Advanced GIS: \url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}

Rauch. C. (2012): Der perfekte Gipfel.  Panorama, 2/2012, S. 112 \url{http://www.alpenverein.de/dav-services/panorama-magazin/dominanz-prominenz-eigenstaendigkeit-eines-berges_aid_11186.html}

Leonhard, W. (2012): Eigenständigkeit von Gipfeln.\url{http://www.thehighrisepages.de/bergtouren/na_orogr.htm}
}
\seealso{
Rauch. C. (2012): Der perfekte Gipfel.  Panorama, 2/2012, S. 112.
Leonhard, W. (2012): Eigenständigkeit von Gipfeln. -

\code{\link{initEnvironGIS}}, \code{\link{calculateDominance}},
\code{\link{calculateProminence}}, \code{\link{calculateEValue}},
\code{\link{makePeak}},
}

