% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateProminence.R
\name{calculateProminence}
\alias{calculateProminence}
\title{Calculates the prominence Value for a given tuple of coordinates and altitude}
\usage{
calculateProminence(x.coord, y.coord, altitude)
}
\arguments{
\item{peaks}{list of all peaks}

\item{altitude}{altitude in meter}

\item{exact.enough}{= vertical excactness of notch altitude in meter}

\item{xcoord}{xcoordinate of current peak (mapunits)}

\item{ycoord}{ycoordinate of current peak (mapunits)}
}
\value{
calculateProminence returns the prominence value of the current peak
}
\description{
Calculates the prominence Value for a given tuple of coordinates and altitude
as derived from DEM d = dominance horizontal distance to the next higher altirude in the sourrounding
}
\examples{
#### Example to calculateProminence

calculateProminence(peaklist.tupel)
}
\author{
Chris Reudenbach
}
\references{
Marburg Open Courseware Advanced GIS: \url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}

Rauch. C. (2012): Der perfekte Gipfel.  Panorama, 2/2012, S. 112 \url{http://www.alpenverein.de/dav-services/panorama-magazin/dominanz-prominenz-eigenstaendigkeit-eines-berges_aid_11186.html}

Leonhard, W. (2012): Eigenständigkeit von Gipfeln.\url{http://www.thehighrisepages.de/bergtouren/na_orogr.htm}
}

