% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distMergePeaks.R
\name{distMergePeaks}
\alias{distMergePeaks}
\title{Trys to merge indenpendent peak positions to provide a reliable join of different data sources}
\usage{
distMergePeaks(dem.peaklist,ext.peaklist)
}
\arguments{
\item{dem.peaklist}{DEM derived peaklist containing coords and altitude}

\item{ext.peaklist}{external peaklist, containing coords name and altitude}
}
\value{
distMergePeaks returns one merged peaklist with the names of the external peaks and the coords and altitude from the DEM derived list
}
\description{
Two list of coordinates and altitudes are compared.
http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description
}
\examples{
#### Example merge two peak lists
      new.peaklist<-distMergePeaks(dem.peaklist,ext.peaklist)
}
\author{
Chris Reudenbach
}
\references{
\url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}
}

