% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setup.R
\name{initEnvironGIS}
\alias{initEnvironGIS}
\title{Function that initializes RSAGA, GRASS GDAL and the R packages}
\usage{
initEnvironGIS(ini.file)
}
\arguments{
\item{DEMfname}{name of used raster DEM (GDAL format)}

\item{ini.file}{name of the session ini.file}
}
\value{
initEnvironGIS initializes the usage of the GIS packages and other utilities
}
\description{
Function that initializes environment pathes, SAGA, GRASS and GDAL support (and the bindings to the corresponding R packages and the R packages *NOTE* you probably have to customize some settings ini file
}
\examples{
#### Example to initialize the enviroment and GIS bindings for use with R
#### uses the ini list from an ini file

ini.example=system.file("demo.ini", package="Rpeak")
dem.example=system.file("test.asc", package="Rpeak")
initEnvironGIS(ini.example,dem.example)
gmeta6()
}
\author{
Chris Reudenbach
}

