% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/initEnvironGIS.R
\name{initEnvironGIS}
\alias{initEnvironGIS}
\title{Function that initializes RSAGA, GRASS GDAL and the R packages}
\usage{
initEnvironGIS(fname.control)
}
\arguments{
\item{fname.control}{name of the session ini.file}

\item{fname.DEM}{name of used raster DEM (GDAL format)}
}
\value{
initEnvironGIS initializes the usage of the GIS packages and other utilities
}
\description{
Function that initializes environment pathes, SAGA, GRASS and GDAL support (and the bindings to the corresponding R packages and the R packages *NOTE* you probably have to customize some settings ini file
}
\details{
For Further Information look at:\link{INI}
}
\examples{
#### Example to initialize the enviroment and GIS bindings for use with R
#### uses the ini list from an ini file

ini.example=system.file("demo.ini", package="perfectPeak")
dem.example=system.file("test.asc", package="perfectPeak")
initEnvironGIS(ini.example,dem.example)
gmeta6()
}
\author{
Chris Reudenbach
}

