% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rpeak.R
\name{Rpeak}
\alias{Rpeak}
\title{Example script that can be used as a wrapper function to start the
perfect peak analysis}
\usage{
Rpeak(ini.file)
}
\arguments{
\item{fname}{character file storing the setup parameters as ini file}

\item{DEMfname}{Digtial Elevation Model has to be a GDAL raster file}
}
\value{
Rpeak returns the complete list as a dataframe of all parameters and results and
generates some output (maps and tables)
}
\description{
Organises all necessary processing steps for calculating the
perfect peak  parameters and generating the output. It  performs preprocessing
and controls the calculations of dominance, prominence, independence (E) value
for a given georeferencend Digital Elevation Model (DEM)\cr
\cr
You can use the function as it is or alternatively use it as skeleton control
script that you cab adapt to your needs.
}
\examples{
#### Example to Rpeak

# It needs a georeferenced DEM as obligatory input. All parameters are read from
# the demo.ini file. You will find some comments in the file.


ini.example=system.file("demo.ini", package="Rpeak")
dem.example=system.file("test.asc", package="Rpeak")
Rpeak(ini.example,dem.example)
}
\author{
Chris Reudenbach
}
\references{
\url{http://moc.environmentalinformatics-marburg.de/doku.php?id=courses:msc:advanced-gis:description}
}
\seealso{
Rauch. C. (2012): Der perfekte Gipfel.  Panorama, 2/2012, S. 112.\cr
\url http://www.alpenverein.de/dav-services/panorama-magazin/dominanz-prominenz-eigenstaendigkeit-eines-berges_aid_11186.html\cr
Leonhard, W. (2012): Eigenständigkeit von Gipfeln. -

\code{\link{initEnvironGIS}}, \code{\link{calculateDominance}},
\code{\link{calculateProminence}},\code{\link{calculateEValue}},
\code{\link{makePeak}},
}

